requirements.txt
kokoro==0.9.4
IPython==8.20.0
soundfile==0.13.1
gradio==4.31.0
numpy==2.3.2
pydantic-ai==0.5.0
python-dotenv==1.0.0